refactor(ipc): split subscriber.py into package with base/sync/async modules
Use Generic[_JsonHandlerT, _RawHandlerT] to deduplicate shared init,
socket lifecycle, route registration, and close into _IpcSubscriberBase.
